pyproject.toml
[project]
name = "ocrvid"
version = "0.1"
description = ""
readme = "README.md"
authors = [{name = ""}]
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
classifiers = [
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "click",
    "google-api-python-client",
    "pyobjc-framework-Vision",
    "opencv-python",
    "pytube",
    "pyserde"
]

[project.urls]
Homepage = "https://github.com/kj-9/ocrvid"
Changelog = "https://github.com/kj-9/ocrvid/releases"
Issues = "https://github.com/kj-9/ocrvid/issues"
CI = "https://github.com/kj-9/ocrvid/actions"

[project.entry-points.console_scripts]
ocrvid = "ocrvid.cli:cli"

[project.optional-dependencies]
test = ["pytest"]
dev = ["pre-commit"]

[tool.setuptools]
packages = ["ocrvid"]

# see also: https://beta.ruff.rs/docs/configuration/#using-pyprojecttoml
[tool.ruff]
select = [
    "E", "F", "I",
    #"D", for now
    "W", "PL", "B", "UP"
]

# Never enforce `E501` (line length violations).
ignore = [
    "E501",
    "PLR2004"# for now
]

# Ignore `E402` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
[tool.ruff.per-file-ignores]
"*__init__.py" = ["F401"]

[tool.ruff.pydocstyle]
convention = "google"
